train_microservice/requirements.txt
google-cloud-storage==1.42.0
numpy==1.21.2
pandas==1.3.2
tensorflow==2.6.0
scikit-learn==0.24.2
